Confidence interval for proportion $$

Confidence intervals for the population proportion $$ for a large n and the population mean $$ under certain conditions.

Usage

ci_prop(p, n, conf.level = 0.95, method = "wald", bounds = "two-sided")

Arguments

bounds

option to choose two-sided confidence intervals, lower or upper bounds. Defaults to two-sided

p:

sample proportion

n:

sample size

conf.level:

confidence level between 0 and 1. Defaults to .95

method:

the method chosen for computing the confidence interval for proportions. Defaults to wald

Details

When n is sufficiently large and p is not too small, the sample proportion p has an approximately Gaussian distribution. In the long run, if repeated samples were taken, the sample proportion would vary around the proportion $$ in the target population.

There are three available methods for computing the binomial proportion confidence interval.

wald: Relies on approximating the distribution of error about a binomially-distributed observation. Typically used when n is sufficiently large and p is not too small.

exact: This method uses the binomial distribution to calculate the coverage probability, which is never less than $1- $. See here for a simple explanation

wilson: Asymetric interval with continuity correction typically employed with small samples and skewed observations. The observed coverage probability is closer to $1- $ than the Wald method.

Examples

ci_prop(.45,100)
#> Error in ci_prop(0.45, 100): could not find function "ci_prop"